backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pyjwt==2.8.0
bcrypt==4.1.3
python-multipart==0.0.9
pydantic==2.12.5
sqlalchemy[asyncio]==2.0.36
greenlet>=3.3.2
asyncpg==0.31.0
aiosqlite==0.21.0
python-dotenv==1.2.1
alembic==1.13.1
pytest==8.4.2
pytest-asyncio==0.26.0
httpx==0.27.2
requests==2.31.0
Pillow==12.1.1
